.github/workflows/requirements.txt
python-telegram-bot==21.0.1
google-generativeai==0.8.3
pillow==10.4.0
python-dotenv==1.0.1